fabs are moved from HomeRoomListFragment to NewHomeDetailFragment
