+++ changelog.d/6765.wip
@@ -1,0 +1,1 @@
+[App Layout] Fabs doesn't go off screen anymore
